Fancy eyecatches, nicer logo added.
Logo uses Praetorian font; see script.rpy. Eyecatches are moderately altered for
code-correctness but basically what Agasa put together; see script.rpy. Also,
having separate methods for eyecatch-taking-one-date and
eyecatch-taking-two-dates is so very behind-the-times; it's not like we're
working in C or something here. (:P) I've replaced all the calls to eyecatch in
the code up to chapter 5, but haven't gone over everything; there might be
instances when it needs changing, so the whole thing should definitely be gone
over by some combination of us.